#13- heroku DB와 연동
헤로쿠 DB와 연동 완료했습니다. 그 과정에서 폴더 위치가 바뀌었습니다.
