+++ .gitignore
@@ -12,6 +12,8 @@ build/
 develop-eggs/
 dist/
 downloads/
+.DS_Store
+.vscode/
 eggs/
 .eggs/
 lib/
@@ -109,6 +111,7 @@ venv/
 ENV/
 env.bak/
 venv.bak/
+my_settings.py
 
 # Spyder project settings
 .spyderproject
