+++ Covid_Project/CovidWeb/CovidWeb/asgi.py
@@ -4,7 +4,7 @@ ASGI config for CovidWeb project.
 It exposes the ASGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/4.1/howto/deployment/asgi/
+https://docs.djangoproject.com/en/3.2/howto/deployment/asgi/
 """
 
 import os
