+++ Covid_Project/CovidWeb/CovidWeb/urls.py
