+++ Covid_Project/CovidWeb/CovidWeb/wsgi.py
@@ -4,7 +4,7 @@ WSGI config for CovidWeb project.
 It exposes the WSGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/4.1/howto/deployment/wsgi/
+https://docs.djangoproject.com/en/3.2/howto/deployment/wsgi/
 """
 
 import os
