+++ Covid_Project/CovidWeb/DB/admin.py
@@ -1,0 +1,13 @@
+# from django.contrib import admin
+
+# # Register your models here.
+# from django.contrib import admin
+# from .models import Post
+
+# admin.site.register(Post)
+# Register your models here.
+from django.contrib import admin
+
+from .models import Symptom
+
+admin.site.register(Symptom)
