+++ Covid_Project/CovidWeb/DB/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SymptomConfig(AppConfig):
+    name = 'DB'
