+++ Covid_Project/CovidWeb/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn CovidWeb.wsgi --log-file -
