+++ Covid_Project/CovidWeb/runtime.txt
@@ -1,0 +1,1 @@
+python-3.9.7
