+++ Covid_Project/CovidWeb/static/main_proto.png
