+++ Covid_Project/CovidWeb/static/pill.png
